Flask delivers the interface correctly
