+++ static/assets/img/Pixhawk_status.png
