+++ static/js/Control/Functionality.js
