+++ static/js/JoystickManagerArduino.txt
