+++ templates/index.html
@@ -11,9 +11,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>TecXotic Controller</title>
 
-    <script src="js/bootstrap.min.js"></script>
-    <link rel="stylesheet" href="css/bootstrap.min.css">
-    <link rel="stylesheet" href="css/style.css">
+    <script src="../static/js/bootstrap.min.js"></script>
+    <link rel="stylesheet" href="../static/css/bootstrap.min.css">
+    <link rel="stylesheet" href="../static/css/style.css">
 
 </head>
 <body class="container" style="min-height: 100vh; min-width: 100vw;  max-height: 100vh; max-width: 100vw;">
@@ -38,16 +38,16 @@
         <div class="col-2">
             <div class="row" style="min-height: 50%; max-height:50%; padding-top:40%;">
                 <div class="col" >
-                    <center> <img id="rov_status" src="assets/img/ROV_status.png" style="width: 100%; "> </center>
+                    <center> <img id="rov_status" src="../static/assets/img/ROV_status.png" style="width: 100%; "> </center>
                 </div>
                 <div class="col" >
-                    <center> <img id="controller_status" src="assets/img/Controller_status.png" style="width: 100%;"> </center>
+                    <center> <img id="controller_status" src="../static/assets/img/Controller_status.png" style="width: 100%;"> </center>
                 </div>
                 <div class="col" >
-                    <center> <img id="pixhawk_status" src="assets/img/Pixhawk_status.png" style="width: 100%;"> </center>
+                    <center> <img id="pixhawk_status" src="../static/assets/img/Pixhawk_status.png" style="width: 100%;"> </center>
                 </div>
                 <div class="col" >
-                    <center> <img id="arm_disarm_status" src="assets/img/ArmDisarm_status.png" style="width: 100%;"> </center>
+                    <center> <img id="arm_disarm_status" src="../static/assets/img/ArmDisarm_status.png" style="width: 100%;"> </center>
                 </div>
             </div>
             <div class="row"  style="min-height: 50%; max-height:50%; padding-top:10%;">
@@ -340,10 +340,10 @@
         import * as THREE from "three/assets/build/three.module.js";
     </script> -->
 
-    <script type="module" src="js/Connection/Connection.js"></script>
-    <script type="module" src="js/Control/Control.js"></script>
-    <script src="js/opencv.js"></script>
-    <script type="module" src="js/Interface.js"></script>
+    <script type="module" src="../static/js/Connection/Connection.js"></script>
+    <script type="module" src="../static/js/Control/Control.js"></script>
+    <script src="../static/js/opencv.js"></script>
+    <script type="module" src="../static/js/Interface.js"></script>
 
 
 </footer>
