--- templates/services/read_config.php
@@ -1,6 +1,0 @@
-<?php
-    $file_path = "../assets/config/config.json";
-    $strJsonFileContents = file_get_contents($file_path);
-    $array = json_decode($strJsonFileContents, true);
-    echo(json_encode($array));
-?>
