--- templates/services/readrov_message.php
@@ -1,15 +1,0 @@
-<?php 
-    require "Connection.php";
-    $data = [];
-    if($pdo!=null){
-        error_log("Connection is not null");
-        $sql = "SELECT * FROM ROV_message ORDER BY id";
-        $stmt = $pdo->prepare($sql);
-        $stmt->execute();
-        while($row = $stmt->fetch(PDO::FETCH_NUM))
-            $data[] = $row;
-    }else{
-        $data = "Connection error";
-    }
-    echo json_encode($data);
-?>
